+++ android/src/main/java/io/ashdavies/commons/compat/CanvasCompat.java
@@ -4,7 +4,6 @@ import android.graphics.Canvas;
 import android.graphics.Paint;
 import android.graphics.RectF;
 import android.os.Build;
-import android.support.annotation.NonNull;
 
 public final class CanvasCompat {
 
@@ -14,15 +13,7 @@ public final class CanvasCompat {
     throw new IllegalStateException("No instances");
   }
 
-  public static void drawRoundRect(
-      Canvas canvas,
-      float left,
-      float top,
-      float right,
-      float bottom,
-      float rx,
-      float ry,
-      @NonNull Paint paint) {
+  public static void drawRoundRect(Canvas canvas, float left, float top, float right, float bottom, float rx, float ry, Paint paint) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
       CanvasCompatApi21.drawRoundRect(canvas, left, top, right, bottom, rx, ry, paint);
     } else {