+++ android/src/main/java/io/ashdavies/commons/exception/NotImplementedException.java
@@ -1,0 +1,4 @@
+package io.ashdavies.commons.exception;
+
+public class NotImplementedException extends RuntimeException {
+}