+++ android/src/main/java/io/ashdavies/commons/factories/FragmentFactory.java
@@ -1,9 +1,0 @@
-package io.ashdavies.commons.factories;
-
-import android.content.Context;
-import android.support.v4.app.Fragment;
-
-public interface FragmentFactory extends InstanceFactory<Fragment> {
-
-  String getTitle(Context context);
-}