+++ android/src/main/java/io/ashdavies/commons/factories/InstanceFactory.java
@@ -1,8 +1,0 @@
-package io.ashdavies.commons.factories;
-
-import android.content.Context;
-
-public interface InstanceFactory<T> {
-
-  T newInstance(Context context);
-}