+++ android/src/main/java/io/ashdavies/commons/presenter/AbstractViewPresenter.java
@@ -1,80 +1,51 @@
 package io.ashdavies.commons.presenter;
 
-import android.content.Context;
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import io.ashdavies.commons.Preconditions;
 import io.ashdavies.commons.view.AbstractView;
-import java.lang.ref.WeakReference;
 
 public abstract class AbstractViewPresenter<View extends AbstractView> implements AbstractView, ViewPresenter<View> {
 
-  private WeakReference<View> reference;
+  @Nullable
+  private View view;
 
   @Override
-  public void onAttach(@NonNull View view) {
-    reference = new WeakReference<>(view);
-    onViewAttached(reference.get());
+  public void onAttach(View view) {
+    this.view = view;
+    onViewAttached();
   }
 
-  @SuppressWarnings("WeakerAccess")
-  protected void onViewAttached(@NonNull View view) {
-    /* no op */
-  }
+  abstract void onViewAttached();
 
-  @Nullable
   protected View getView() {
-    return reference.get();
+    if (view == null) {
+      throw new ViewNotAttachedException();
+    }
+
+    return view;
   }
 
   protected boolean isAttached() {
-    return reference != null && reference.get() != null;
+    return view != null;
   }
 
   @Override
   public void onDetach() {
-    reference.clear();
+    this.view = null;
     onViewDetached();
   }
 
-  @SuppressWarnings("WeakerAccess")
-  protected void onViewDetached() {
-    /* no op */
+  abstract void onViewDetached();
+
+  @Override
+  public void onProgress(boolean progressing) {
+    getView().onProgress(progressing);
   }
 
   @Override
-  public Context getContext() {
-    return Preconditions.notNull(getView(), "View not attached").getContext();
+  public void onError(Throwable throwable) {
+    getView().onError(throwable);
   }
 
-  @Override
-  public void showProgress() {
-    Preconditions.require(getView(), new Preconditions.Action<View>() {
-      @Override
-      public void run(View view) {
-        view.showProgress();
-      }
-    });
-  }
-
-  @Override
-  public void hideProgress() {
-    Preconditions.require(getView(), new Preconditions.Action<View>() {
-      @Override
-      public void run(View view) {
-        view.hideProgress();
-      }
-    });
-  }
-
-  @Override
-  public void onError(@NonNull final Throwable throwable) {
-    Preconditions.require(getView(), new Preconditions.Action<View>() {
-      @Override
-      public void run(View view) {
-        view.onError(throwable);
-        view.hideProgress();
-      }
-    });
+  private static class ViewNotAttachedException extends RuntimeException {
   }
 }