+++ android/src/main/java/io/ashdavies/commons/presenter/ViewPresenter.java
@@ -1,12 +1,10 @@
 package io.ashdavies.commons.presenter;
 
-import android.support.annotation.NonNull;
-
 public interface ViewPresenter<View> {
 
-  void onAttach(@NonNull View view);
+  void onAttach(View view);
 
   void onDetach();
 
-  void onError(@NonNull Throwable throwable);
+  void onError(Throwable throwable);
 }