+++ android/src/main/java/io/ashdavies/commons/view/CollectionView.java
@@ -2,7 +2,7 @@ package io.ashdavies.commons.view;
 
 import java.util.Collection;
 
-public interface ListView<T> extends AbstractView {
+public interface CollectionView<T> extends AbstractView {
 
   void add(T item);