+++ android/src/main/java/io/ashdavies/commons/view/ProgressView.java
@@ -1,0 +1,6 @@
+package io.ashdavies.commons.view;
+
+public interface ProgressView {
+
+  void onProgress(boolean progressing);
+}